grid/list view switching via single css class (#26)
i've taken some often-used-together foundation framework css classes and
put their content into a new class together. this saves a few bytes per
journal entry and probably makes rendering faster, but creates more work
for us devs when e.g. updating foundation.

there is still (a lot of) cruft left to whittle down, so i'm not going
to make a PR out of this yet. but this should be a usable branch for
performance testing of the new view switch