requires config.php changes: bootstrap include mechanism changed
this commit is pretty much all over the place, refactoring how
config.php and sys/bootstrap.php are loaded and/or load each other and
the i18n stuff. also, a number of places that ingest network supplied
input have gotten their sanity upgraded (#32).

this should probably go into the 0.2 milestone since
existing installations need to compare their config.php
to config-default.php and merge the changes.

there is some performance won with this, since the config object is passed
to class constructors that require it now and not parsed from the file more than once.

this should probably be on its own branch, but it relies on the changes
introduced in view_refactoring a lot :S i kinda got carried away
there. gonna try and re-apply the network input related stuff to develop
once view_refactoring is merged in. just had to put this somewhere :)